#171 Free channels when we're done playing sound on them.
This is, I think, a major contributing factor to sound spatialization
messing up in busy areas. Basically, we were only freeing channels with
S_SpatializeChannel deemed they were out of range. Now, for non-looped
samples, we explicitly free the channel as soon as the sample is done
playing. This opens up the channel to another voice, whereas previously
it would have just skipped the sound.