#171 Big ol' s_mix refactor (#194)
* #171, this is close, but still needs work.

* Code review fixes.

* #171 Nailed it.

Mix_HaltChannel won’t call your Mix_ChannelFinished callback if the
channel isn’t actually playing. So in that case, we have to free the
channel ourselves. Otherwise, we fill up s_env.channels with zombies.

* #171 Increase MAX_CHANNELS

On a map with a lot of ambient sounds (Torn Glory), running around and
firing the machinegun can consume almost 30 channels. Doesn’t leave a
lot of room for 3 players firing machineguns. Increasing this to 128
channels.

* #171 Cosmetics.

* #171 Cosmetics.