Allow boosting health past 100 on players
Implemented with a timer, rather then the item think approach. (I didn't see a good way to have small items decrease the total without stepping on the mega). It's up to a mapper to not have to many mega's.

Other thought that was unimplemented was to decrease health faster from higher full health. May revisit later.