#235 Fix plat start / end sounds.
The bug here was that the entity origin is not in any PHS when the
entity is at the bottom of its movement. So firing a normal entity
sound would appear to fail. The solution is to emit these sounds as
positioned sounds from the center of the entity’s abs_max’s.