Feature/60hz (#245)
Something is strangely wrong with this PR. I rebased off master after @Paril's particles merge, and yet those commits are showing here as new. I'm going to merge this so that we can make sense of it all and iron things out in master. It's stable and functional, tho there are likely small regressions to sort out.

* Intermediate commit towards 60hz server tick.

* Fix prediction.

* Use QUETOO_TICK_RATE as cl_max_fps lower bounds.

* Enable debugging output for whatever compiler you're using.

* Formatting.

* Fix a crash in pmove where we could blow out planes array.

* Add Q3's unique plane detection to Pm_SlideMove.

* Fix high DPI for 2d matrix in MVC

* Documentation improvements around lerp.

* Add style target to top level Makefile to invoke astyle.

* Feature/particles (#246)

* Add VectorLengthSquared func, use it for element depths
Expose CullBox, going to see if I can do some magic with that
Sort particles only with each other by image ID.

* Add CullSphere function, which should be faster to compare than boxes for extremely simple stuff
Cull particles. This seems to help immensely on weather maps.
Undo qsort changes, they don't work since qsort is only one pass it seems.

* Two sort passes, one for particles.

* Fixed up the second sorting pass, it now finds particle ranges then sorts them. Apparently qsort did not like doing a full pass otherwise.
Coronas are now particles. The only part of this that I find to be ugly is that r_particle_t is a pointer type so I can't just use a static instance of one and push it into the list (see r_draw_bsp_lights; had to add an r_particle_t to the bsp light structure so that it had one to pull from).

* Update particles during the sort pass to prevent a second loop.

* Atlaseses

* More atlas features - specifically, implemented in cgame finally. Some particles, such as scrolling beams, cannot use the atlas for obvious reasons (at least not without some complicated shader work that isn't really worth implementing atm).

* Quake 3 frustum calculations actually work with widescreen.
Added cull pass/fail to cl_show_renderer_stats
Beams cull from center with length(start - end) radius

* Fix Xcode and Autotools builds.

* Fix first person beams to actually come from gun.

* Some nice thin beam-like sparks for bullets. jdolan pls check if good.

* This seems to work

* Fixed some docs in cgame

* Revert an early change since it's no longer necessary

* Fix up requested changes. Also cleaning up some old stuff.

* Generate -> Compile

* Typo whoops

* TE_SPARKS

* Intermediate commit towards 60hz server tick.

* Fix prediction.

* Use QUETOO_TICK_RATE as cl_max_fps lower bounds.

* Enable debugging output for whatever compiler you're using.

* Formatting.

* Fix a crash in pmove where we could blow out planes array.

* Add Q3's unique plane detection to Pm_SlideMove.

* Documentation improvements around lerp.

* Add style target to top level Makefile to invoke astyle.