Rough cut of moving kick effect interpolation client side.
There are some issues with this. Namely, I’ve moved the kick_angles out
of player_state_t, and am sending them over as a pair, only when they
are set by a server-side event. This is more efficient, but I fear what
I may have broken..

For starters, the kick angles do not affect your command angles. So
your aim stays stationary even when your view is kicked. I’m not
completely opposed to this, but it’s a little weird.