OpenAL sound driver (#469)
* This is a semi-'sperimental branch right now, but basically the idea is to replace SDL_mixer with OpenAL + SDL_sound.
Rationale: SDL_mixer is garbage at playing audio. Even with just this basic OpenAL setup working, it's already sounding so much better. SDL_sound is used to load the audio,  In this commit, nearly everything is working as expected, however:
- music is not implemented
- ogg is not supported on Windows yet
- Windows-related setup for OpenAL SDK home directory. Need to also have it copy Win64 to x64, otherwise the props don't work right
- fade-out of S_PLAY_FRAME sounds doesn't work yet. need to investigate this one.
- recommend that you use soft_oal instead of the default OpenAL driver as it's much better. The default OAL has a weird thing where some sounds' attack is very soft, whereas with soft_oal it sounds identical to the old sound subsystem. We can bundle this with the Windows version probably.

* Compile SDL_sound with OGG support. Woo.
Add velocity calculation, not using it but it was fun to play with.

* Music streaming! Yay.
Notes:
 - this is done on the main thread right now, so decoding will be interrupted by map changes, etc. Plan to move it to its own thread.
 - Buffer count & size is cvar, so if users are having issues with streaming they can adjust it on the fly.

* Threaded music processing. Woooooo, crispy.

* Sorta-kinda fix threading for music.
Currently, having threads >= 0 and then setting to threads -1 will cause a softlock because of trying to close threads, including the sound one, which runs forever. Need a way to separate promises (one-off tasks that perform an operation and return a value) from threads that run the entire length of the program.

* Use direct SDL threads to avoid Thread_ deadlock

* Update OS X, Xcode and Autotools builds for OpenAL and SDL_sound.

* Use SDL_RWops for sound loading.

* Don't leak the entire music file :)

* Use libsndfile instead

* Support OGG/FLAC
Always use file rate for music. Can't use resample because it doesn't work on stereo yet.

* Decided to remove stereo->mono code, I think it's best if we allow stereo sounds (since there are some cases where stereo sounds are used, like countdown oggs) but just not bother spatializing them. Stereo sounds can be played just fine so long as they are global or on the player.

Resampling stereo sounds needs to be fixed up, though, and I heard some popping in music playback but unsure if that was just me or not.

* Fix issue with AL_LOOPING
Global error-checking

* Comment function.

* Pass right channel # for AL

* Update Xcode and Autotools builds for libsndfile.

* Fix a couple casting warnings.

* We only need SDL_rwops.h in sound now.

* Fix crash in S_Shutdown on OS X: Make NULL the current context before destroying ours.

* Update mingw dependencies script for SDL_mixer -> OpenAL and friends.

* The code might be a bit overkill, but, stereoresampling is here! If the allocations cause an issue we can use a single static buffer for the splitting/resampling or something, sorta like what the music system does.
Added a global Mem_Check which is just used for programmatic debugging. Used it to track down a heap overrun.
Adjusted wording ("samples" should not be used when function returns # of frames. The wording is confusing, libsndname uses "item" vs "frame"; "frame" = "item * num_channels", "item" = "sample")

* Throw warning on stereo spatialized samples

* Document s_env, and get rid of "initialized" since we can now check the context.

* Proper stereo resampling code, yay.
Fix warning in cm_material.c
NearestMultiple now handles 0 properly
(Play with s_rate 5000: Duke Nukem mode! music sounds awesome too, haha)
Remove unnecessary #define for WinSock

* Pitch testing. Also somehow these didn't get pushed.

* Ah shit

* Fix clipping issues with OGG (float overflow in the ogg files caused a wraparound, SFC_SET_SCALE_FLOAT_INT_READ causes it to clamp properly)

* Fix MinGW OpenAL libs.