Introduce r_lightmap_t and r_lightmap_media_t, and attempt to calculate uniform luxel lightmap coordinates in the renderer (not working).
r_lightmap_t is an analog for lightmap_t in quemap. It contains per-surface lightmap information, such as mins, maxs and a transform matrix.

r_lightmap_media_t is an aggregate for layered, atlased lightmap and deluxemap images, as well as the stainmap framebuffer render target.

This drastically cleans up all lightmap management. Even if the texture coordinates calculations need to be reverted, I want to keep these structs.