Smoothstep functions like in GLSL. Touch spotlight falloff.
light->theta is in radians so the cone dotproduct needs to be compared to cos(light->theta) and not just light->theta. The default value of LIGHT_CONE is 22.5 degrees (1/8th a circle), but this is the angle as bisected by the light normal. Which means the total angle is 45 degrees (1/4th a circle). I think the latter is more intuitive for mappers?