Proof of concept for BSP ordered transparency.
The idea is to walk the tree from back to front, drawing particles that are behind each alpha blended BSP plane before drawing the blended elements on that plane. The plane is transformed into view space and passed to the shader for culling. This way, we retain all of our batching optimizations, and use a geometry shader to preempt rendering of particles that are in front of the plane in question.