Remove epsilon from Cm_BoxOnPlaneSide, like Quake3 does.
This fixes (at least) many misses for Cm_BoxLeafnums. An interesting unit test that deserves more research here is to check Cm_BoxLeafnums(leaf->mins, leaf->maxs) for each leaf. The function _should_ return at least every leaf for itself, but currently it does not. This commit gets it working for the majority of leafs, but not all.