Kill off more int16_t in favor of larger types, including protocol number.
The code is just so much easier to deal with if you rarely have to think about integer type sizes. Quake3 (and now our code to a large extent) uses 32 bit ints for everything. We should only use 16 bit types when there's a clear, easy to remember reason to -- such as network IO.