Use Vec3_BoxIntersect on draw elements / node to ensure draw elements are added in the correct order.
Because multiple nodes reference the same plane, it's critical that we only add a draw element when it resides within the currently visited node. This ensures proper depth sorting of draw elements, while still allowing them to be aggregated by plane.

There are probably corner cases where this will fall down. But I tested this on Edge where the cobwebs overlap the water, and it works. I also tested it on Sandbox, which has contrived depth sorting challenges, and it looked good, with stable ordering and believable visuals.