Simplify depth pass elements by storing them on the BSP.
Also, don't depend on the cvar for allocating or freeing the buffer. This had a leak in it if I loaded the map with depth pass, disabled depth pass, and unloaded the map. Additionally, this buffer is not that large, so always allocate it so that r_restart is not required to toggle depth pass.