Remove Free function on inline BSP model to fix media crash.
The blend_elements pointer array is allocated in R_LoadBspInlineModel, not in R_SetupBspInlineModel (media conversion). Because the inline model is link-malloc'ed to the BSP model, it is freed when the BSP model is freed. Therefore, we should free blend elements then, too, and remove the free function from inline model media.