WIP faces reference brush sides.
Something that always bothered me about the Quake BSP format is that faces had no relationship to the brush sides that created them. This patch fixes that. Faces no longer duplicate plane, texinfo and contents. Instead, they point to the original brush side in the map file that created them. This is not only a more efficient format, but it allows us to do more deterministic and WYSIWYG face merging for flares, etc.

There are issues with blended faces in this patch so far. That's why it's WIP. I think it has to do with portal / plane sidedness.

This patch will require no effort from map authors, but it will require a rebake of all maps (BSP format change).