Remove timers 😬
These are tripping up Xcode to a point where I can't even think straight 😂 Indentation is broken on every line. Syntax highlighting only works half of the time. Breakpoints don't work where they should. It's infuriating.

I think we've gotten the profiling information we were looking for with these. I'd prefer to use more targeted tooling if we need to do this again, but we can always revert this commit if we discover we want them back.