BREAKING CHANGE reintroduce face merging.
This saves 10-30% of triangles on the maps I've tested. Fewer tris means marginally better performance, but fewer lighting and texturing artifacts. This algorithm also finds a few more faces to merge than the original.

The breaking change here is to add the plane reference back on the face. This is because, for translucent brushes, portals may create faces on either side of each brush side (the inner walls of glass, or the surface of water, for example). So it is not safe to assume that the face plane is always the brush side plane.