The last piece of the puzzle.
Don't double-dot diffuse light, as this diminishes it and requires "brightness" levels of 3+. The shader is the best place to apply this (with the normalmap). So instead of applying the dot product in quemap to nerf the light, scale the direction by the light intensity. This finally produces bumpmaps that look just like our dynamic lights.